Add Several Improvements to the Project
Made some changes to the  and  files, which's fixing a bug with the Spawning Area of enemies, Made serveral changes to , it's hard to cover all of it in this description, so please... read the source file for more information

As well as improving the Simple Debugger by alot! from fixing the Overlaping text rendering issue, to adding functions to the de
bugger which allow it to, Toggle visibility of the Debugger, and increase/decrease text scale of the Debugger.

Please read the changes made in this commit for more context.